### 2025-02-21 #7834 @editorialbot
[PRE REVIEW]: rmcmc: Robust Markov chain Monte Carlo methods in R
**Submitting author:** @matt-graham (<a href="http://orcid.org/0000-0001-9104-7960">Matthew Graham</a>)
**Repository:** https://github.com/UCL/rmcmc
**Branch with paper.md** (empty if default branch): mmg/joss-paper
**Version:** v0.1.1
