### 2025-02-21 #7834 @matt-graham
>  if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

[larryshamalama](https://github.com/larryshamalama), [vankesteren](https://github.com/vankesteren), [robmoss](https://github.com/robmoss), [njtierney](https://github.com/njtierney) and [martinmodrak](https://github.com/martinmodrak) all seem from their reviewer profiles like they could be good fits in terms of a combination of expertise in R and Bayesian statistics / Monte Carlo methods
