### 2025-02-21 b003557
Adding missing DOIs to references
