+++ joss/references.bib
@@ -5,7 +5,8 @@
   journal = {Statistics and Computing},
   publisher = {Springer},
   volume = 18,
-  pages = {343--373}
+  pages = {343--373},
+  doi = {10.1007/s11222-008-9110-y}
 }
 @article{barker1965monte,
   title = {{M}onte {C}arlo calculations of the radial distribution functions for a proton electron plasma},
@@ -15,7 +16,8 @@
   publisher = {CSIRO Publishing},
   volume = 18,
   number = 2,
-  pages = {119--134}
+  pages = {119--134},
+  doi = {10.1071/ph650119}
 }
 @article{besag1994comments,
   title = {Comments on {"Representations of knowledge in complex systems"} by {U. Grenander} and {MI Miller}},
@@ -35,7 +37,8 @@
   publisher = {Int Statistical Institute},
   volume = 19,
   number = {5a},
-  pages = {1501--1534}
+  pages = {1501--1534},
+  doi = {10.3150/12-bej414}
 }
 @article{betancourt2017conceptual,
   title = {A conceptual introduction to {Hamiltonian Monte Carlo}},
@@ -48,7 +51,8 @@
   author = {Paul-Christian Bürkner and Jonah Gabry and Matthew Kay and Aki Vehtari},
   year = 2024,
   url = {https://mc-stan.org/posterior/},
-  note = {R package version 1.6.0}
+  note = {R package version 1.6.0},
+  doi = {10.32614/cran.package.posterior}
 }
 @article{carpenter2017stan,
   title = {{Stan}: A probabilistic programming language},
@@ -97,14 +101,16 @@
   pages = {97--109},
   issn = {00063444, 14643510},
   url = {http://www.jstor.org/stable/2334940},
-  urldate = {2025-02-17}
+  urldate = {2025-02-17},
+  doi = {10.1093/oso/9780198509936.003.0015}
 }
 @manual{helske2021ramcmc,
   title = {{ramcmc}: Robust Adaptive {M}etropolis Algorithm},
   author = {Jouni Helske},
   year = 2021,
   url = {https://github.com/helske/ramcmc},
-  note = {R package version 0.1.2}
+  note = {R package version 0.1.2},
+  doi = {10.32614/cran.package.ramcmc}
 }
 @inproceedings{hird2020fresh,
   title = {A fresh take on {‘Barker dynamics’} for {MCMC}},
@@ -112,7 +118,8 @@
   year = 2020,
   booktitle = {International Conference on Monte Carlo and Quasi-Monte Carlo Methods in Scientific Computing},
   pages = {169--184},
-  organization = {Springer}
+  organization = {Springer},
+  doi = {10.1007/978-3-030-98319-2_8}
 }
 @article{hoffman2014no,
   title = {The {No-U-Turn} sampler: adaptively setting path lengths in {Hamiltonian Monte Carlo}.},
@@ -129,7 +136,8 @@
   year = 2012,
   journal = {The Annals of Statistics},
   publisher = {JSTOR},
-  pages = {3050--3076}
+  pages = {3050--3076},
+  doi = {10.1214/12-aos1048}
 }
 @article{livingstone2022barker,
   title = {The {B}arker proposal: combining robustness and efficiency in gradient-based {MCMC}},
@@ -161,7 +169,8 @@
   publisher = {American Institute of Physics},
   volume = 21,
   number = 6,
-  pages = {1087--1092}
+  pages = {1087--1092},
+  doi = {10.1007/s12045-022-1419-x}
 }
 @incollection{neal2011mcmc,
   title = {{MCMC} Using {Hamiltonian} Dynamics},
@@ -169,7 +178,8 @@
   year = 2011,
[…]
