### 2025-02-21 #7834 @matt-graham
> ```
> ❌ MISSING DOIs
>
> - 10.1007/s11222-008-9110-y may be a valid DOI for title: A tutorial on adaptive MCMC
> - 10.1071/ph650119 may be a valid DOI for title: Monte Carlo calculations of the radial distributio...
> - 10.3150/12-bej414 may be a valid DOI for title: Optimal tuning of the Hybrid Monte-Carlo algorithm
> - 10.32614/cran.package.posterior may be a valid DOI for title: posterior: Tools for Working with Posterior Distri...
> - 10.1093/oso/9780198509936.003.0015 may be a valid DOI for title: Monte Carlo Sampling Methods Using Markov Chains a...
> - 10.32614/cran.package.ramcmc may be a valid DOI for title: ramcmc: Robust Adaptive Metropolis Algorithm
> - 10.1007/978-3-030-98319-2_8 may be a valid DOI for title: A fresh take on ‘Barker dynamics’ for MCMC
> - 10.1214/12-aos1048 may be a valid DOI for title: Variable transformation to obtain geometric ergodi...
> - 10.1007/s12045-022-1419-x may be a valid DOI for title: Equation of state calculations by fast computing m...
> - 10.1201/b10905-6 may be a valid DOI for title: MCMC Using Hamiltonian Dynamics
> - 10.2139/ssrn.912637 may be a valid DOI for title: Primal-dual subgradient methods for convex problem...
> - 10.2307/3318418 may be a valid DOI for title: Exponential Convergence of Langevin Distributions ...
> - 10.1111/1467-9868.00123 may be a valid DOI for title: Optimal scaling of discrete approximations to Lang...
> - 10.1214/ss/1015346320 may be a valid DOI for title: Optimal scaling for various Metropolis–Hastings al...
> - 10.1063/1.436415 may be a valid DOI for title: Brownian dynamics as smart Monte Carlo simulation
> - 10.1093/biomet/asac056 may be a valid DOI for title: Optimal design of the Barker proposal and other lo...
> ```

Just to note these should now be fixed (all seemed to valid DOIs for referenced items when I checked them) on current head of `mmg/joss-paper` branch - I don't think I have permissions to run the `@editorialbot check references` command though as looks like [this is reserved to editors](https://joss.readthedocs.io/en/latest/editorial_bot.html#check-references).
