### 2025-02-24 348eb2b
Adding JOSS submission status badge
