### 2025-02-24 34ee475
Avoid badge link wrapping
