### 2025-02-24 ac4a7e6
Adding JOSS submission status badge (#76)
* Adding JOSS submission status badge

* Avoid badge link wrapping
