+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   adjusting-noise-distribution: adjusting-noise-distribution.html
   barker-proposal: barker-proposal.html
-last_built: 2025-02-19T09:48Z
+last_built: 2025-02-24T11:12Z
 urls:
   reference: http://github-pages.ucl.ac.uk/rmcmc/reference
   article: http://github-pages.ucl.ac.uk/rmcmc/articles
