### 2025-03-03 50bd844
Factor out function for checking initial state
