### 2025-03-03 4764a17
Factor out functions for initializing target distribution and trace function
