### 2025-03-03 c8cea53
Simplify state threading logic
