### 2025-03-03 12a8572
Factor out functions in chain loop
