### 2025-04-29 #7834 @skanwal
Hello [@fabian-s](https://reviewers.joss.theoj.org/reviewers/343) - are you available to review this submission?
