### 2025-05-13 #7834 @skanwal
Hello [@thomaspinder](https://github.com/thomaspinder) - are you available to review this submission for JOSS?
