### 2025-06-04 #7834 @skanwal
Hello @njtierney  - are you available to review this submission for JOSS?
