### 2025-07-16 UCL/rmcmc#80 @AntoineSoetewey
Is the license OSI approved?
Dear @matt-graham, the license, as written in the GitHub repo, is as follows:

```
YEAR: 2025
COPYRIGHT HOLDER: rmcmc authors
```

is this a [OSI approved](https://opensource.org/licenses/alphabetical) software license?

Reference: https://github.com/openjournals/joss-reviews/issues/8594
