### 2025-07-16 UCL/rmcmc#81 @AntoineSoetewey
Add section title "References" in the paper
In the paper: Can you please add a section title "References" before the references (after the acknowledgement)?

Ref: https://github.com/openjournals/joss-reviews/issues/8594
