### 2025-07-16 UCL/rmcmc#82 @AntoineSoetewey
Typo in paper
"this can significantly improvement sampling efficiency in some cases." shouldn't be "this can significantly improve sampling efficiency in some cases."?

Ref: https://github.com/openjournals/joss-reviews/issues/8594
