### 2025-07-16 UCL/rmcmc#83 @AntoineSoetewey
Make summary section of paper more accessible
The Summary section of the paper provides a good overview of the software’s functionality and purpose, particularly for readers familiar with MCMC methods. However, to make this section more accessible and to better satisfy the checklist criterion, I suggest revising it slightly to ensure a clearer description of the **high-level functionality and purpose** of the software for a **diverse, non-specialist audience**.

Ref: https://github.com/openjournals/joss-reviews/issues/8594
