### 2025-07-16 UCL/rmcmc#84 @AntoineSoetewey
Simplify dense or technical sentences in paper
The paper is well written and clearly structured. To improve accessibility for a broader audience, could you please consider simplifying a few dense or technical sentences in the paper?

Ref: https://github.com/openjournals/joss-reviews/issues/8594
