### 2025-07-16 UCL/rmcmc#85 @AntoineSoetewey
Add real-world example
The documentation provides clear examples of how to use the package. However, to better demonstrate the practical value of the software, it would be helpful to include an example based on a real-world analysis problem and/or using actual data. This would make it easier for users to see how `rmcmc` fits into applied workflows.

Ref: https://github.com/openjournals/joss-reviews/issues/8594
