### 2025-07-16 UCL/rmcmc#80 @matt-graham
Hi @AntoineSoetewey. The package is [MIT licensed](https://github.com/UCL/rmcmc/blob/main/LICENSE.md) which is an OSI approved software license. The `LICENSE` file you posted the contents of is in the format required for CRAN packages [which does not allow copies of standard licenses to be included in the package](https://usethis.r-lib.org/reference/licenses.html#details). Having both `LICENSE` and `LICENSE.md` files is [a standard approach to deal with this in R packages](https://r-pkgs.org/license.html#key-files) - see for example the [usethis repository](https://github.com/r-lib/usethis/) or [testthat](https://github.com/r-lib/testthat).
