### 2025-07-16 634c632
Add references section title
Resolves #81
