### 2025-07-16 4be0659
Correct typo improvement -> improve
Resolves #82
