### 2025-07-16 UCL/rmcmc#81 @matt-graham
Thanks for spotting this - now fixed by 634c632
