### 2025-07-17 UCL/rmcmc#84 @AntoineSoetewey
For instance I find the sentences in lines 34-38, 42-46, 53-57 and 79-82 quite long, don't you think?

I’m not suggesting cutting important details, just that breaking some of these up into shorter sentences might improve readability without sacrificing clarity.
