### 2025-07-17 UCL/rmcmc#86 @AntoineSoetewey
Missing background information in the paper
Again, the paper is well-written but I feel that some background information, in particular on MCMC, is missing. This background information could help non-specialists better understand the topic and grasp why the package is useful.

Ref: https://github.com/openjournals/joss-reviews/issues/8594
