### 2025-07-17 UCL/rmcmc#85 @matt-graham
Thanks @AntoineSoetewey, this is a good suggestion I'll have a look at adding an additional vignette article illustrating using the package on a real-world example.
