### 2025-07-24 4f5f5cb
Simplifying and shortening sentences in paper
Relates to #84
