+++ joss/paper.md
@@ -30,7 +30,7 @@ Generating samples from a probability distribution
 is a common requirement in many disciplines.
 In Bayesian inference, for example, the distribution of interest is the posterior over parameters of a model,
 given data assumed to be generated from the model.
-Expectations with respect to the posterior can be easily estimated using samples,
+Expectations with respect to the posterior can be estimated using samples,
 allowing computation of posterior means, variances and other quantities of interest.
 _Markov chain Monte Carlo_ (MCMC) methods
 are a general class of algorithms for approximately sampling from probability distributions.
@@ -38,57 +38,64 @@ are a general class of algorithms for approximately sampling from probability di
 `rmcmc` is an R package providing implementations of MCMC methods for sampling from distributions on $\mathbb{R}^d$ for $d \geq 1$.
 It provides a general purpose implementation of the Barker proposal [@livingstone2022barker],
 a gradient-based MCMC algorithm inspired by the Barker accept-reject rule [@barker1965monte].
-It also provides implementations of other MCMC algorithms based on random walk, Langevin and Hamiltonian dynamics,
-and a flexible interface for performing adaptive MCMC
+`rmcmc` also provides implementations of other MCMC algorithms based on random walk, Langevin and Hamiltonian dynamics.
+It has a flexible interface for performing adaptive MCMC
 so that algorithmic tuning parameters can be learned in a bespoke manner [@haario2001adaptive;@andrieu2008tutorial].
 The key function provided by the package is `sample_chain`,
-which allows sampling a Markov chain with a user-specified stationary distribution.
+which samples a Markov chain with a user-specified stationary distribution.
 The chain is sampled by generating proposals
 and accepting or rejecting them using the Metropolis--Hastings [@metropolis1953equation;@hastings1970monte] algorithm.
-During an initial warm-up stage, the parameters of the proposal distribution can be adapted,
-with schemes available to both:
-tune the scale of the proposals by coercing the average acceptance rate to a target value;
-tune the shape of the proposals to match covariance estimates under the target distribution.
+During an initial warm-up stage, the parameters of the proposal distribution can be adapted.
+Schemes are available to both tune:
+
+- the scale of the proposals by coercing the average acceptance rate to a target value,
+- the shape of the proposals to match covariance estimates under the target distribution.
 
 # Statement of need
 
 For target distributions with smooth density functions,
-gradient-based MCMC methods can provide significantly improved sampling efficiency
-over simpler schemes such as _random-walk Metropolis_ (RWM),
-particularly in high-dimensional settings [@roberts1996exponential;@roberts1998optimal;@beskos2013optimal].
-For schemes such as _Metropolis adjusted Langevin algorithm_ (MALA) [@rossky1978brownian;@besag1994comments]
+gradient-based MCMC methods can provide improved sampling efficiency
+over simpler schemes such as _random-walk Metropolis_ (RWM)
+in high-dimensional settings [@roberts1996exponential;@roberts1998optimal;@beskos2013optimal].
+For methods such as _Metropolis adjusted Langevin algorithm_ (MALA) [@rossky1978brownian;@besag1994comments]
 and _Hamiltonian Monte Carlo_ (HMC)  [@duane1987hybrid;@neal2011mcmc],
-this improved efficiency can, however, come at the cost of decreased robustness to tuning of the algorithm's parameters,
+this improved efficiency comes with a cost.
+Specifically these methods exhibit decreased robustness to tuning of the algorithm's parameters,
 compared to non-gradient based methods such as RWM [@livingstone2022barker].
-The Barker proposal provides a middle road,
-offering similar efficiency in high-dimensional settings as other gradient-based methods such as MALA [@vogrinc2023optimal],
[…]
