### 2025-07-24 UCL/rmcmc#84 @matt-graham
Hi @AntoineSoetewey, apologies for the delay in getting back to you. I have now split up / simplified the sentences you specifically flagged and also shortened some other long sentences (based on a relatively arbitrary threshold of >30 words) in 4f5f5cb. Let me know if this resolves your concerns about readability here!
