### 2025-08-02 UCL/rmcmc#87 @AntoineSoetewey
Open issues
Dear @matt-graham,

I noticed that there are some open issues, tagged as enhancement (in addition to the ones I created for the review). Do you plan to address these in future updates of the package?

Ref: https://github.com/openjournals/joss-reviews/issues/8594#issuecomment-3076080744
