### 2025-08-04 6863d8b
Add article with example of interfacing with Stan model
