+++ .github/workflows/pre-commit.yaml
@@ -21,7 +21,7 @@ jobs:
       - name: Install R dependencies
         uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::pkgdown, any::styler, github::roualdes/bridgestan/R@v2.5.0, local::.
+          extra-packages: any::pkgdown, any::styler, github::roualdes/bridgestan/R@v2.6.2, local::.
 
       - name: Set up Python
         uses: actions/setup-python@v5
