+++ .github/workflows/test-coverage.yaml
@@ -37,7 +37,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::covr, any::xml2, github::roualdes/bridgestan/R@v2.5.0
+          extra-packages: any::covr, any::xml2, github::roualdes/bridgestan/R@v2.6.2
           needs: coverage
 
       - name: Test coverage
