+++ .github/workflows/lint.yaml
@@ -36,7 +36,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::lintr, any::cyclocomp, github::roualdes/bridgestan/R@v2.5.0, local::.
+          extra-packages: any::lintr, any::cyclocomp, github::roualdes/bridgestan/R@v2.6.2, local::.
           needs: lint
 
       - name: Lint
