+++ .github/workflows/pkgdown.yaml
@@ -66,7 +66,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::pkgdown, github::roualdes/bridgestan/R@v2.5.0, local::.
+          extra-packages: any::pkgdown, github::roualdes/bridgestan/R@v2.6.2, local::.
           needs: website
 
       - name: Build site
