+++ .github/workflows/test-coverage.yaml
@@ -37,8 +37,9 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::covr, any::xml2, github::roualdes/bridgestan/R@v2.6.2
-          needs: coverage
+          needs: |
+            bridgestan
+            coverage
 
       - name: Test coverage
         run: |
