+++ DESCRIPTION
@@ -46,6 +46,20 @@ Suggests:
     rmarkdown,
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
+Config/Needs/bridgestan:
+    github::roualdes/bridgestan/R@v2.6.2
+Config/Needs/check:
+    any::rcmdcheck
+Config/Needs/coverage:
+     any::covr,
+     any::xml2
+Config/Needs/lint:
+    any::lintr,
+    any::cyclocomp
+Config/Needs/style:
+    any::styler
+Config/Needs/website:
+    any::pkgdown
 URL: https://github.com/UCL/rmcmc, http://github-pages.ucl.ac.uk/rmcmc/
 BugReports: https://github.com/UCL/rmcmc/issues
 Imports: 
