+++ .github/workflows/pkgdown.yaml
@@ -66,8 +66,10 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::pkgdown, github::roualdes/bridgestan/R@v2.6.2, local::.
-          needs: website
+          extra-packages: local::.
+          needs: |
+            bridgestan
+            website
 
       - name: Build site
         run: pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
