### 2025-08-07 cc63e1c
Try install bridgestan from r-multiverse src package
