### 2025-08-07 97ce325
Add lint dependencies to pre-commit workflow spec
