### 2025-08-07 e7c9fc6
Add bridgestan to lint workflow dependencies
