### 2025-08-07 ce44a7e
Update pre-commit config
