### 2025-08-07 9fd8365
Expand details of variables and reflow sentences
