### 2025-08-07 b578a18
Use pre-commit action rather than manual approach
