+++ .github/workflows/pre-commit.yaml
@@ -13,28 +13,10 @@ jobs:
       - name: Checkout source
         uses: actions/checkout@v4
 
-      - name: Set up R
-        uses: r-lib/actions/setup-r@v2
-        with:
-          use-public-rspm: true
-
-      - name: Install R dependencies
-        uses: r-lib/actions/setup-r-dependencies@v2
-        with:
-          extra-packages: local::.
-          needs: |
-            bridgestan
-            lint
-            style
-            website
-
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
           python-version: "3.x"
 
-      - name: Install Python dependencies
-        run: python -m pip install pre-commit
-
       - name: Run pre-commit
-        run: pre-commit run --all-files --color always --verbose
+        uses: pre-commit/action@v3.0.1
