### 2025-08-07 668b1a6
Restore r-setup step
