+++ .github/workflows/pre-commit.yaml
@@ -13,6 +13,11 @@ jobs:
       - name: Checkout source
         uses: actions/checkout@v4
 
+      - name: Set up R
+        uses: r-lib/actions/setup-r@v2
+        with:
+          use-public-rspm: true
+
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
