### 2025-08-07 60af597
Remove lintr pre-commit hook due to cyclocomp dep issue
