### 2025-08-07 42bb9fa
Try installing additional dependencies via setup-r-dependencies
