### 2025-08-07 9a995cb
Don't install local package in setup-r-dependencies
