+++ .github/workflows/pre-commit.yaml
@@ -21,6 +21,7 @@ jobs:
       - name: Install R dependencies
         uses: r-lib/actions/setup-r-dependencies@v2
         with:
+          packages: any::sessioninfo
           needs: |
             lint
             style
