+++ .github/workflows/pre-commit.yaml
@@ -21,8 +21,8 @@ jobs:
       - name: Install R dependencies
         uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          packages: any::sessioninfo
           needs: |
+            bridgestan
             lint
             style
             website
