### 2025-08-07 4852998
Update to latest version of BridgeStan in workflows + update pre-commit config (#91)
* Update to latest version of BridgeStan in workflows

* Use Config/Needs to avoid replication

* Try install bridgestan from r-multiverse src package

* Adding explicit package name

* Add lint dependencies to pre-commit workflow spec

* Add bridgestan to lint workflow dependencies

* Update pre-commit config

* Update pre-commit version

* Use pre-commit action rather than manual approach

* Restore r-setup step

* Remove lintr pre-commit hook due to cyclocomp dep issue

* Try installing additional dependenc
[…]
