+++ DESCRIPTION
@@ -61,7 +61,8 @@ Config/Needs/style:
 Config/Needs/website:
     any::pkgdown,
     ggplot2,
-    bayesplot
+    bayesplot,
+    rjson
 URL: https://github.com/UCL/rmcmc, http://github-pages.ucl.ac.uk/rmcmc/
 BugReports: https://github.com/UCL/rmcmc/issues
 Imports: 
