### 2025-08-07 9cc3c3d
Add article with example of interfacing with Stan model (#90)
* Add article with example of interfacing with Stan model

* Expand details of variables and reflow sentences

* Add rjson to website dependencies

* Add hexbin to website dependencies

* Add gridExtra package to website dependencies
