+++ vignettes/articles/references.bib
@@ -1,0 +1,19 @@
+@article{hodgkin1952currents,
+  title={Currents carried by sodium and potassium ions through the membrane of the giant axon of {L}oligo},
+  author={Hodgkin, Allan L and Huxley, Andrew F},
+  journal={The Journal of Physiology},
+  volume={116},
+  number={4},
+  pages={449},
+  year={1952}
+}
+@article{daly2015hodgkin,
+  title={{Hodgkin--Huxley} revisited: reparametrization and identifiability analysis of the classic action potential model with approximate {B}ayesian methods},
+  author={Daly, Aidan C and Gavaghan, David J and Holmes, Chris and Cooper, Jonathan},
+  journal={Royal Society open science},
+  volume={2},
+  number={12},
+  pages={150499},
+  year={2015},
+  publisher={The Royal Society Publishing}
+}
