+++ vignettes/articles/interfacing-with-stan-models.Rmd
@@ -1,0 +1,343 @@
+---
+title: "Interfacing with Stan models via BridgeStan"
+bibliography: references.bib
+link-citations: true
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+The `rmcmc` package includes support for using the probabilistic programming language Stan 
+to define the target distribution for inference via the 
+[R interface of the excellent BridgeStan package](https://roualdes.us/bridgestan/latest/languages/r.html).
+This article illustrates how to the Stan modelling syntax to define a model to use with `rmcmc` 
+using an example model and data from computational neuroscience.
+
+
+```{r setup}
+library(rmcmc)
+```
+
+## Hodgkin-Huxley model for action potential generation
+
+We will use as an example a mathematical model describing how action potentials in neurons are generated
+and how we might fit that model to data. 
+The [model of interest](https://en.wikipedia.org/wiki/Hodgkin%E2%80%93Huxley_model) 
+was described by @hodgkin1952currents based on experimental investigations with squid giant axons.
+They received the Nobel Prize in Physiology or Medicine for their work in 1963.
+We will fit the model here to a digitisation of the data the original experiments from @daly2015hodgkin.
+
+The [cell membrane voltage $V_m$](https://en.wikipedia.org/wiki/Membrane_potential) 
+is modelled as being controlled by [voltage-gated ion channels](https://en.wikipedia.org/wiki/Voltage-gated_ion_channel),
+with governing [ordinary differential equation (ODE)](https://en.wikipedia.org/wiki/Ordinary_differential_equation)
+
+$$
+ C_m \frac{\mathrm{d}V_m}{\mathrm{d}t} = g_K (E_K - V_m) + g_{Na}(E_{Na} - V_m) + g_L(E_l - V_m) - I.
+$$
+
+where 
+
+- $C_m$ is the membrane capacitance per unit area, 
+- $g_K$ is the potassium channel conductance, 
+- $E_k$ the potassium channel reversal potential,
+- $g_{Na}$ is the sodium channel conductance,
+- $E_{Na}$ is the sodium channel reversal potential,
+- $g_L$ is the leakage conductance,
+- $E_L$ is the leakage reversal potential,
+- and $I$ is the membrane current per unit area.
+
+Importantly the potassium and sodium channel conductances, $g_K$ and $g_{Na}$, 
+are themselves time and voltage dependent variables. 
+In Hodgkin and Huxley's experiments they used an ingenious [experimental protocol](https://en.wikipedia.org/wiki/Voltage_clamp)
+to allow measuring the evolution of the cell membrane's conductance over time for specific ion channels
+while clamped to a specific depolarization voltage. 
+By varying the extracellular ion concentrations, 
+they were able to isolate the effects of the potassium and sodium channels. 
+Here we will concentrate on the model and data specifically for $g_{K}$, the potassium channel conductance. 
+
+The potassium channel conductance is modelled as
+
+$$g_{\textrm{K}}(t, v, \theta) = \bar{g}_{\textrm{K}} n(t, v, \theta)^4,$$
+
+where
+
+- $\bar{g}_K$ is the maximum potassium channel conductance, 
+- $n$ is a time and voltage dependent subunit activation probability in $[0, 1]$, 
+- $v = V_{\text{rest}} - V_m$ is the depolarization from the resting potential $V_{\text{rest}}$ in mV 
+- and $\theta = (\bar{g}_K, k_{\alpha,1}, k_{\alpha, 2}, k_{\alpha, 3}, k_{\beta, 1}, k_{\beta,2})$ are a set of model parameters.
+
+The potassium channel subunit activation $n$ is governed by a linear ODE
+
[…]
