+++ pkgdown.yml
@@ -4,7 +4,8 @@ pkgdown_sha: ~
 articles:
   adjusting-noise-distribution: adjusting-noise-distribution.html
   barker-proposal: barker-proposal.html
-last_built: 2025-08-07T18:51Z
+  articles/interfacing-with-stan-models: interfacing-with-stan-models.html
+last_built: 2025-08-07T20:00Z
 urls:
   reference: http://github-pages.ucl.ac.uk/rmcmc/reference
   article: http://github-pages.ucl.ac.uk/rmcmc/articles
