### 2025-08-08 4c6d172
Manually install and invoke pre-commit in Actions workflow
