### 2025-08-08 ac93b04
Fix inclusion of generated quantities in Stan model output
