### 2025-08-08 a7fb850
Document seed argument
