### 2025-08-08 cb23895
Add link to interfacing with Stan article to README (#95)
* Add link to interfacing with Stan article to README

* Fix typo and indentation in article
