+++ README.Rmd
@@ -41,10 +41,12 @@ During an initial warm-up stage, the parameters of the proposal distribution can
 with adapters available to both: 
 tune the scale of the proposals by coercing the average acceptance rate to a target value; 
 tune the shape of the proposals to match covariance estimates under the target distribution. 
-As well as the default Barker proposal, the package also provides implementations of alternative proposal distributions, such as (Gaussian) random walk and Langevin proposals. 
+As well as the default Barker proposal, the package also provides implementations of alternative proposal distributions,
+such as (Gaussian) random walk and Langevin proposals. 
 Optionally, if [BridgeStan's R interface](https://roualdes.github.io/bridgestan/latest/languages/r.html),
 available [on GitHub](https://github.com/roualdes/bridgestan), is installed, 
-then BridgeStan can be used to specify the target distribution to sample from.
+[then BridgeStan can be used to specify the target distribution to sample
+from](https://github-pages.ucl.ac.uk/rmcmc/articles/interfacing-with-stan-models.html).
 
 ## Installation
 
@@ -63,7 +65,7 @@ devtools::install_github("UCL/rmcmc")
 
 ## Examples
 
-The snippet belows shows a basic example of using the package to generate samples from a normal target distribution with random scales. 
+The snippet below shows a basic example of using the package to generate samples from a normal target distribution with random scales. 
 Adapters are used to tune the proposal scale to achieve a target average acceptance probability,
 and to tune the proposal shape with per-dimension scale factors based on online estimates of the target distribution variances.
 
