+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   adjusting-noise-distribution: adjusting-noise-distribution.html
   barker-proposal: barker-proposal.html
   articles/interfacing-with-stan-models: interfacing-with-stan-models.html
-last_built: 2025-08-07T20:00Z
+last_built: 2025-08-08T09:12Z
 urls:
   reference: http://github-pages.ucl.ac.uk/rmcmc/reference
   article: http://github-pages.ucl.ac.uk/rmcmc/articles
