### 2025-08-08 6e2bd6e
Fix inclusion of generated quantities in Stan model output (#94)
* Fix inclusion of generated quantities in Stan model output

* Document seed argument

* Reflow line and explicit arg names
