### 2025-08-08 21e0405
Fix error messages for invalid adapter type arguments
