### 2025-08-08 UCL/rmcmc#83 @matt-graham
Hi @AntoineSoetewey. Just to follow up on this issue, similar to our discussion in #84, could you give some specific comments about which parts of the summary you feel need rewriting? I was aware of the requirement that the summary is a _description of the high-level functionality and purpose of the software for a diverse, non-specialist audience_ when writing the summary, so from my perspective I've already tried to fulfill this!

The first paragraph was intended to give some context to readers who are not familiar with MCMC methods and my hope was this it gives a reasonably accessible description of what MCMC does (generate approximate samples from a probability distribution) without diving in to technical details. If it is the use of statistical terms like 'Expectations', 'means', 'variances' then I could potentially remove or replace some of these, for example 'Integrals' rather than 'Expectations', 'averages' rather than 'means'?

The second paragraph does contain a number of technical terms when describing the functionality of the package, but I'm not sure how to avoid this here, as understanding the features offered by the package probably does require some familiarity with MCMC methods.
