### 2025-08-08 UCL/rmcmc#83 @AntoineSoetewey
Thanks for clarifying. You can actually disregard this issue; making the Summary section more accessible than it already is would likely require adding extra explanatory text (I believe), which would not be appropriate given the word limit imposed by JOSS.
