+++ .github/workflows/pre-commit.yaml
@@ -28,7 +28,7 @@ jobs:
             website
 
       - name: Set up Python
-        uses: actions/setup-python@v5
+        uses: actions/setup-python@v6
         with:
           python-version: "3.x"
 
