### 2025-09-10 #8594 @larryshamalama
Hi @matt-graham, @skanwal , apologies for the significant delay in my review. The paper is very well written and, following @AntoineSoetewey 's raised points, the codebase (documentation, vignette, tests, etc.) looks great too.

Given the focus on Barker proposal (gradient info + Metropolis), my small nitpick would be to elaborate even more on the algorithm in the README, e.g. why both on their own can be unsuitable (particularly in higher dimensional settings?) but their combination can work well. The JOSS paper makes reference to this in the first paragraph of "Statement of need", but I understood a bit more when skimming the JRSSB paper. I leave this up to the package authors. That being said, I did not have time to go through all the details in the Barker proposal paper given my current schedule.

In the interest of time, I would like to recommend this JOSS paper for acceptance; all items in my checklist have been checked. Please let me know if anything else is needed from my end.
