### 2025-09-14 60874ab
Update to bridgestan v0.7.0 in workflows (#99)
