### 2025-10-17 3c332d3
Update NEWS in preparation for v0.1.2 release
