+++ DESCRIPTION
@@ -29,7 +29,7 @@ Description: Functions for simulating Markov chains using the Barker proposal
     distribution. As well as the default Barker proposal, the package also
     provides implementations of alternative proposal distributions, such as
     (Gaussian) random walk and Langevin proposals. Optionally, if 'BridgeStan's
-    R interface <https://roualdes.github.io/bridgestan/latest/languages/r.html>,
+    R interface <https://roualdes.us/bridgestan/latest/languages/r.html>,
     available on GitHub <https://github.com/roualdes/bridgestan>, is installed,
     then 'BridgeStan' can be used to specify the target distribution to sample
     from.
