### 2025-10-17 1905f05
Updating CRAN comments for v0.1.2 release (#103)
