+++ NEWS.md
@@ -1,4 +1,4 @@
-# rmcmc (development version)
+# rmcmc 0.1.2
 
 * In `scale_adapter()` and `shape_adapter()`, setting `algorithm` and `type`
   arguments to invalid values now gives informative error message (#88).
