### 2025-10-21 7bca8aa
Update references.bib with R manual DOI
