+++ joss/references.bib
@@ -215,6 +215,7 @@
   title = {{R}: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
   year = 2024,
+  doi = {10.32614/r.manuals},
   address = {Vienna, Austria},
   url = {https://www.R-project.org/},
   organization = {R Foundation for Statistical Computing}
