### 2025-11-01 #8594 @crvernon
@matt-graham - this submission is in great shape.  I just need you to fix the following before I accept for publication:

In you paper:

- [ ] LINE 177:  The "r" in "An r package..." should be capitalized.  You can do this in your bib file by placing any characters you wish to maintain formatting of in curly brackets.

Let me know when this has been addressed and I'll accept.  Thanks.
