### 2025-11-03 45e2e69
Fix capitalisation in nimbleHMC article title
