+++ joss/references.bib
@@ -287,7 +287,7 @@
   note = {R package version 2.32.6}
 }
 @article{turek2024nimblehmc,
-  title = {{nimbleHMC}: An R package for {Hamiltonian Monte Carlo} sampling in nimble},
+  title = {{nimbleHMC}: An {R} package for {Hamiltonian Monte Carlo} sampling in nimble},
   author = {Daniel Turek and Perry de Valpine and Christopher J. Paciorek},
   year = 2024,
   journal = {Journal of Open Source Software},
