### 2025-11-03 #8594 @matt-graham
> [@matt-graham](https://github.com/matt-graham) - this submission is in great shape. I just need you to fix the following before I accept for publication:
>
> In you paper:
>
>     * [ ]  LINE 177:  The "r" in "An r package..." should be capitalized.  You can do this in your bib file by placing any characters you wish to maintain formatting of in curly brackets.
>
>
> Let me know when this has been addressed and I'll accept. Thanks.

Thanks @crvernon - [this is now fixed](https://github.com/UCL/rmcmc/pull/71/commits/45e2e69338bc0b942ad5c03a7abd9204d9673f7e).
